feat(urlprefix): add url strings for each page in the project #87
