+++ App.tsx
@@ -8,7 +8,6 @@ import {
 } from "react-native";
 import { Provider as PaperProvider } from "react-native-paper";
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
-import * as Linking from "expo-linking";
 import { store } from "./Utils/store";
 import { Provider } from "react-redux";
 import RootNavigation from "./Navigation/RootNavigation";
@@ -16,22 +15,10 @@ import { RootStackParamList } from "./Utils/types";
 
 // Uncomment ReduxStateTest to test various state actions and reducers
 // import ReduxStateTest from "./Components/ReduxStateTest";
-const urlPrefix = Linking.createURL("/");
+
 const RootStack = createNativeStackNavigator<RootStackParamList>();
 
 export default function App() {
-  const config = {
-    screens: {
-      Login: "login",
-      Overview: "overview",
-    },
-  };
-
-  const linking = {
-    prefixes: [urlPrefix],
-    config,
-  };
-
   return (
     <Provider store={store}>
       <PaperProvider>
